fix(build): move wagmi to features/wallet-connect per FSD
Fixes SSR build error on /create page caused by RainbowKit initialization
during static generation.

Root cause: Barrel-file side effects - importing CREATOR_STORE_KEY from
@/shared/config triggered wagmiConfig initialization (getDefaultConfig)
which requires WALLETCONNECT_PROJECT_ID at module load time.

Solution:
- Move wagmi.ts: shared/config → features/wallet-connect/config
- Move Web3Provider: shared/providers → features/wallet-connect/providers
- Remove wagmiConfig from barrel exports to prevent side effects
- Preserve lazy loading via dynamic() in app/lazy-web3-provider.tsx

FSD compliance: Web3 configuration belongs to wallet-connect feature,
not shared infrastructure layer.
