+++ .gitignore
@@ -37,6 +37,10 @@ node_modules
 .next/
 out/
 
+# Generated by next-sitemap
+public/sitemap*.xml
+public/robots.txt
+
 # Production
 build/
 dist/
