--- public/sitemap.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://voidpay.xyz/apple-icon</loc><lastmod>2025-12-24T18:37:21.274Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://voidpay.xyz/icon</loc><lastmod>2025-12-24T18:37:21.276Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://voidpay.xyz</loc><lastmod>2025-12-24T18:37:21.276Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>https://voidpay.xyz/create</loc><lastmod>2025-12-24T18:37:21.276Z</lastmod><changefreq>weekly</changefreq><priority>0.9</priority></url>
-<url><loc>https://voidpay.xyz/history</loc><lastmod>2025-12-24T18:37:21.276Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-</urlset>
