+++ src/features/wallet-connect/providers/web3-scope.tsx
